fix(widgets): satisfy quality gate on the Dataset & Mapping widget
- Extract the dataset-status block into DatasetStatusSection so
  DatasetWidget drops under the 200-line cap (221 -> 141).
- Use the --weight-semi token instead of a raw font-weight in the
  connection-editor styles.

[allow-copyright]
