+++ apps/web/src/ui/domains/widgets/navigation/DatasetWidget.tsx
@@ -7,7 +7,6 @@
  */
 
 import { useState, useEffect, useCallback, useRef, useMemo } from 'react';
-import type { CSSProperties } from 'react';
 import { useGameUIStore } from '../../../../stores/game-ui-store';
 import { useNavigationOverlayStore } from '../../../../stores/navigation-overlay-store';
 import { getDungeonName } from '@shared/game/data/screens/game-values';
@@ -18,25 +17,14 @@ import { connectionIssues } from './connection-issues';
 import { useConnectionAudit } from './useConnectionAudit';
 import { useRealTransitions } from './useRealTransitions';
 import { ConnectionAuditSection } from './ConnectionAuditSection';
+import { DatasetStatusSection } from './DatasetStatusSection';
 import { ScreenEditorDialog } from './ScreenEditorDialog';
 import { ConnectionEditorDialog } from './ConnectionEditorDialog';
-import { Box, Text, StatusBadge, Button } from '../../../design-system/primitives';
+import { Box } from '../../../design-system/primitives';
 import { useScreenDetection } from './hooks';
 import type { ReviewStatus, ReviewData } from './dataset-widget-types';
 import { S } from './dataset-widget-styles';
 import { StatusRow } from './StatusRow';
-import { DatasetStatusPill } from './DatasetStatusPill';
-
-// Static inline-style literals (dynamic/conditional styles stay inline).
-const IL: Record<string, CSSProperties> = {
-  statusHead: { display: 'flex', alignItems: 'center', gap: 6, marginBottom: 4 },
-  warnText: { color: 'var(--c-warning)', fontSize: 10 },
-  corrBox: { padding: '3px 6px', marginTop: 2, borderRadius: 'var(--r-sm)', background: 'var(--c-warning-soft)', border: '1px solid var(--c-warning-soft)' },
-  corrTitle: { fontSize: 9, color: 'var(--c-warning)', fontWeight: 600, marginBottom: 2 },
-  corrItem: { fontSize: 10, color: 'var(--c-text-dim)', lineHeight: '14px' },
-  corrField: { color: 'var(--c-info)' },
-  btnRow: { display: 'flex', gap: 4, marginTop: 4 },
-};
 
 const DatasetWidgetContent = () => {
   const { overworldScreenIndex, roomIndex, isIndoors, isDarkWorld, palaceIndex } = useGameUIStore(s => s.map);
@@ -150,85 +138,15 @@ const DatasetWidgetContent = () => {
   return (
     <Box style={S.root}>
       {/* ═══ DATASET STATUS ═══ */}
-      <Box style={S.section}>
-        <Box style={IL.statusHead}>
-          <Box style={S.sectionTitle}>Dataset</Box>
-          <DatasetStatusPill
-            background={screenStatus.status === 'mapped' ? '#1a3a1a' : screenStatus.status === 'incomplete' ? '#3a3a1a' : '#3a1a1a'}
-            color={screenStatus.status === 'mapped' ? 'var(--c-green)' : screenStatus.status === 'incomplete' ? 'var(--c-warning)' : 'var(--c-danger)'}
-          >
-            {screenStatus.status === 'mapped' ? '✓ Screen' : screenStatus.status === 'incomplete' ? '⚠ Screen' : '✗ Screen'}
-          </DatasetStatusPill>
-          <DatasetStatusPill
-            background={connStatus.status === 'complete' ? '#1a3a1a' : connStatus.status === 'partial' ? '#3a3a1a' : '#2a2a2a'}
-            color={connStatus.status === 'complete' ? 'var(--c-green)' : connStatus.status === 'partial' ? 'var(--c-warning)' : 'var(--c-text-muted)'}
-          >
-            {connStatus.status === 'complete' ? '✓ Conns' : connStatus.status === 'partial' ? `⚠ ${connStatus.missingCount} missing` : '— Conns'}
-          </DatasetStatusPill>
-        </Box>
-        <Box style={S.infoBox}>
-          <Box style={S.infoRow}>
-            <Text style={S.infoLabel}>Screen</Text>
-            <Text style={{ color: screenStatus.screen ? 'var(--c-green-bright)' : 'var(--c-danger)' }}>
-              {screenStatus.screen ? screenStatus.screen.id : 'Not mapped'}
-            </Text>
-          </Box>
-          {detectionResult && (
-            <Box style={S.infoRow}>
-              <Text style={S.infoLabel}>Match</Text>
-              <Text style={{ color: detectionResult.method === 'exact' || detectionResult.method === 'overworld' ? 'var(--c-green)' : detectionResult.method === 'entrance' ? 'var(--c-info)' : 'var(--c-warning)' }}>
-                {detectionResult.method}
-              </Text>
-            </Box>
-          )}
-          {screenStatus.screen && (
-            <Box style={S.infoRow}>
-              <Text style={S.infoLabel}>Name</Text>
-              <Text>{screenStatus.screen.name}</Text>
-            </Box>
-          )}
-          {screenStatus.screen && (
-            <Box style={S.infoRow}>
-              <Text style={S.infoLabel}>Status</Text>
-              <StatusBadge status={screenStatus.screen.status} />
-            </Box>
-          )}
-          {screenStatus.screen && !screenStatus.screen.variant && detectionResult?.method === 'cave-ambiguous' && progressInfo && (
-            <Box style={S.infoRow}>
-              <Text style={S.infoLabel}>⚠️</Text>
-              <Text style={IL.warnText}>Default entry — no variant for "{progressInfo.label}"</Text>
-            </Box>
-          )}
-          {screenStatus.issues.length > 0 && (
-            <Box style={S.infoRow}>
-              <Text style={S.infoLabel}>Issues</Text>
-              <Text style={IL.warnText}>{screenStatus.issues.join(', ')}</Text>
-            </Box>
-          )}
-          {screenStatus.corrections.length > 0 && (
-            <Box style={IL.corrBox}>
-              <Box style={IL.corrTitle}>⚠ Suggested Corrections</Box>
-              {screenStatus.corrections.map((c, i) => (
-                <Box key={i} style={IL.corrItem}>
-                  <Text style={IL.corrField}>{c.field}</Text>: {c.message}
-                </Box>
-              ))}
-            </Box>
-          )}
-          <Box style={S.infoRow}>
-            <Text style={S.infoLabel}>Connections</Text>
-            <Text>{connStatus.existingConnections.length} in dataset{connStatus.missingCount > 0 ? `, ${connStatus.missingCount} detected not mapped` : ''}{incompleteConnCount > 0 ? `, ${incompleteConnCount} incomplete` : ''}</Text>
-          </Box>
-        </Box>
-        <Box style={IL.btnRow}>
-          <Button variant="tertiary" size="sm" onClick={() => setScreenEditorOpen(true)}>
-            ✏️ Edit Screen
-          </Button>
-          <Button variant="tertiary" size="sm" onClick={() => setConnEditorOpen(true)}>
-            ✏️ Edit Connections
-          </Button>
-        </Box>
-      </Box>
+      <DatasetStatusSection
+        screenStatus={screenStatus}
+        connStatus={connStatus}
+        detectionResult={detectionResult}
+        progressInfo={progressInfo}
+        incompleteConnCount={incompleteConnCount}
+        onEditScreen={() => setScreenEditorOpen(true)}
+        onEditConnections={() => setConnEditorOpen(true)}
+      />
 
       {/* ═══ CONNECTION AUDIT ═══ */}
       <ConnectionAuditSection badFindings={audit.badFindings} addFindings={audit.addFindings} />
